Formatting + Features Added
Re-did the entire GUI and the ability to set an IP address and subnet mask and an interface description.

I also did some formatting issues to make the code more readable and malleable.

Next, I'll be adding the option to configure either a L2 or L3 interface to fully utilize the new IP address feature.